+++ src/Adnd.Server/Migrations/20260802194025_NPCStatusAndLastSeen.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Adnd.Server.Migrations
+{
+    /// <inheritdoc />
+    public partial class NPCStatusAndLastSeen : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<DateTimeOffset>(
+                name: "LastSeenAt",
+                table: "NPCs",
+                type: "timestamp with time zone",
+                nullable: true);
+
+            migrationBuilder.AddColumn<int>(
+                name: "Status",
+                table: "NPCs",
+                type: "integer",
+                nullable: false,
+                defaultValue: 0);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "LastSeenAt",
+                table: "NPCs");
+
+            migrationBuilder.DropColumn(
+                name: "Status",
+                table: "NPCs");
+        }
+    }
+}